+++ src/Exception/InvalidToken.php
@@ -2,7 +2,7 @@
 
 namespace Firebase\Auth\Token\Exception;
 
-use Exception;
+use Throwable;
 use Lcobucci\JWT\Token;
 
 class InvalidToken extends \InvalidArgumentException
@@ -12,14 +12,14 @@ class InvalidToken extends \InvalidArgumentException
      */
     private $token;
 
-    public function __construct(Token $token, $message = '', $code = 0, Exception $previous = null)
+    public function __construct(Token $token, $message = '', $code = 0, Throwable $previous = null)
     {
         parent::__construct($message, $code, $previous);
 
         $this->token = $token;
     }
 
-    public function getToken()
+    public function getToken(): Token
     {
         return $this->token;
     }